Pre-PR cleanups
