+++ frontend/src/pages/requests/RequestsPage.tsx
@@ -1,7 +1,6 @@
 import {
   RequestWhere,
   SampleWhere,
-  useFindSamplesByInputValueQuery,
   useRequestsListLazyQuery,
 } from "../../generated/graphql";
 import { useState } from "react";
@@ -9,16 +8,13 @@ import {
   RequestsListColumns,
   SampleDetailsColumns,
   defaultEditableColDef,
+  getMetadataFromSamples,
   sampleFilter,
 } from "../../shared/helpers";
 import RecordsList from "../../components/RecordsList";
 import { useParams } from "react-router-dom";
 import PageHeader from "../../shared/components/PageHeader";
 import { parseSearchQueries } from "../../utils/parseSearchQueries";
-import {
-  getAllSamplesFromQueryData,
-  getMetadataFromSamples,
-} from "../../shared/utils";
 
 function requestFilterWhereVariables(uniqueQueries: string[]): RequestWhere[] {
   if (uniqueQueries.length > 1) {
@@ -86,8 +82,6 @@ export default function RequestsPage() {
         conditionBuilder={requestFilterWhereVariables}
         sampleQueryParamFieldName={sampleQueryParamFieldName}
         sampleQueryParamValue={params[sampleQueryParamFieldName]}
-        useSampleRecordsQuery={useFindSamplesByInputValueQuery}
-        getSamplesFromQueryData={getAllSamplesFromQueryData}
         sampleDefaultColDef={defaultEditableColDef}
         getRowData={getMetadataFromSamples}
         sampleColDefs={SampleDetailsColumns}
